% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imzMLreader.R
\name{check_imzMLUUID}
\alias{check_imzMLUUID}
\title{check_imzMLUUID.}
\usage{
check_imzMLUUID(imzML_file, fixibd_uuid = F, xmlParserResult = NULL)
}
\arguments{
\item{imzML_file}{path to an imzML file.}

\item{fixibd_uuid}{set to TRUE to overwrite the ibd file uuid with the imzML uuid.}

\item{xmlParserResult}{result of the CimzMLParse() if available}
}
\value{
TRUE if the UUID is OK.
export
}
\description{
Check and fix the UUID of invalid imzML files.
}
