% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{estimateTolerance}
\alias{estimateTolerance}
\title{estimates the tolerance in the imzML file data.}
\usage{
estimateTolerance(fileName, params)
}
\arguments{
\item{fileName:}{Absolute paths of the file.}

\item{params:}{(optional)
"SNR": signal-to-noise ratio (by defect=10)
"noiseMethod": method for estimating noise (by defect="estnoise_mad").
nThreads: number of threads for parallel processing (by default maxCores-1)
imzMLChecksum: if the binary file checksum must be verified, it can be disabled for convenience with really big files.
fixBrokenUUID: set to FALSE by default to automatically fix an uuid mismatch between the ibd and the imzML files (a warning message will be raised).}
}
\value{
a matrix with the tolerance estimate for ten mass segments.
}
\description{
estimates the tolerance in the imzML file data.
}
