% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getBasic}
\alias{getBasic}
\title{returns generic information about centroids located in file _peakMatrix.bin.}
\usage{
getBasic(data)
}
\arguments{
\item{data}{-> a list from getPeakMatriz()}
}
\value{
a list:
mass: Vector with the masses associated with each column of peakMatrix.
intensity: vector of average intensities of all pixels associated to centroid
tolerance: Vector of tolerances achieved for this centroid (ppm)
pixelsSupport: Vector with the number of pixels in each column with non-zero intensity.
pixelsSample: Vector with the number of pixels in each of the samples. In the peakMatrix and coordinates they appear in the same order.
}
\description{
returns generic information about centroids located in file _peakMatrix.bin.
}
