% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getCentroid()}
\alias{getCentroid()}
\alias{getCentroid}
\title{returns a column information of the peak matrix:}
\usage{
getCentroid(data, mass, sample = 1)
}
\arguments{
\item{data}{-> a list from getPeakMatriz()}

\item{mass}{-> reference to the desired initial column of the peak matrix (Da).}

\item{sample}{-> just download the pixels from this sample.
if sample=0, all sample coordinates are returned}
}
\value{
a list:
intensity: vector of intesities
mass: mass associated with the column of peakMatrix.
tolerance: Tolerance achieved for this centroid (ppm)
pixelsSupport: number of pixels in column with non-zero intensity.
}
\description{
returns a column information of the peak matrix:
}
