% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getGaussiansFromSpectrum}
\alias{getGaussiansFromSpectrum}
\title{Form Gaussians over each peak of the given spectrum.}
\usage{
getGaussiansFromSpectrum(
  intensity,
  mz,
  lowMass = 0,
  highMass = 0,
  SNR = 3,
  noiseMethod = "estnoise_mad"
)
}
\arguments{
\item{Intensity:}{vector with value associated with each mass point.}

\item{mz:}{vector with mass/charge information}

\item{lowMass:}{lower mass to consider}

\item{highMass:}{higher mass to consider}

\item{SNR:}{signal-to-noise ratio}

\item{noiseMethod:}{method for estimating noise (estnoise_diff, estnoise_sd, estnoise_mad).}
}
\value{
List: gaussians, mass intensity, SNR, noise
gaussians: matrix with parameters for each Gaussian
mass: vector with the masses of the raw spectrum
intensity: intensity associated with each mass of the raw spectrum
SNR: signal-to-noise ratio associated with each mass of the raw spectrum.
noise: noise estimation
}
\description{
Form Gaussians over each peak of the given spectrum.
}
