% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getIntensityFromFile()}
\alias{getIntensityFromFile()}
\alias{getIntensityFromFile}
\title{returns a column whit pixels intensities from the peak matrix:}
\usage{
getIntensityFromFile(file, column)
}
\arguments{
\item{file}{-> file name with peak matrix (tmpPeakMatrix.bin)}

\item{column}{-> desired column (first = 1)}
}
\value{
a vector with intensity of each pixel.
}
\description{
returns a column whit pixels intensities from the peak matrix:
}
