% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getMassVectorFromFile()}
\alias{getMassVectorFromFile()}
\alias{getMassVectorFromFile}
\title{returns a vector with all the masses in peak matrix}
\usage{
getMassVectorFromFile(file)
}
\arguments{
\item{file}{-> file name with peak matrix (tmpPeakMatrix.bin)}
}
\value{
mass vector
}
\description{
returns a vector with all the masses in peak matrix
}
