% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getPeakMatrix}
\alias{getPeakMatrix}
\title{obtains the peak matrix from imzML files.}
\usage{
getPeakMatrix(
  data_file,
  params,
  lowMass = 0,
  highMass = 0,
  pxList = c(0),
  nThreads = 0,
  imzMLChecksum = F,
  fixBrokenUUID = F
)
}
\arguments{
\item{params}{"massResolution": desired mass resolution.
"SNR": signal-to-noise ratio (by defect=1)
"noiseMethod": method for estimating noise (by defect="estnoise_mad").
"minPixelsSupport": minimum percentage of pixels that must support an ion for it to be considered (by defect=1).
"linkedPeaks": two peaks are considered linked if they are closer than the given standard deviation (by defect=3).}

\item{data_file:}{If multiple samples need to be analyzed, the absolute paths for each sample must be included in a .txt file, line by line.
If only one sample is being analyzed, it can be provided directly as an argument. Only files in imzML format are recognized.}

\item{lowMass:}{lower mass to consider}

\item{highMass:}{higher mass to consider}

\item{pxList:}{list of pixels. First pixel=1. By default everyone.}

\item{nThreads:}{number of threads for parallel processing (by default maxCores-1)}

\item{imzMLChecksum:}{if the binary file checksum must be verified, it can be disabled for convenice with really big files.}

\item{fixBrokenUUID:}{set to FALSE by default to automatically fix an uuid mismatch between the ibd and the imzML files (a warning message will be raised).}
}
\value{
a list:
peakMatrix: Matrix of peak (centroids) rows = pixels, columns = intensity of each pixel.
mass: Vector with the masses associated with each column of peakMatrix.
pixelsSupport: Vector with the number of pixels in each column with non-zero intensity.
}
\description{
obtains the peak matrix from imzML files.
}
