% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getPeakMatrixFromFile}
\alias{getPeakMatrixFromFile}
\title{return the peak matrix into file}
\usage{
getPeakMatrixFromFile(file)
}
\arguments{
\item{file}{-> file name with peak matrix (tmpPeakMatrix.bin)}
}
\value{
a list:
peakMatrix: Matrix of peak (centroids) rows = pixels, columns = intensity of each pixel.
mass: Vector with the masses associated with each column of peakMatrix.
massResolution: mass resolution associated with each mass.
pixelsSupport: Vector with the number of pixels in each column with non-zero intensity.
}
\description{
return the peak matrix into file
}
