% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getPixelsCoordinates}
\alias{getPixelsCoordinates}
\title{returns a matrix with the coordinates of all pixels (X/Y).
If there are multiple samples, they appear sequentially; that is, the matrix has as many rows
as the cumulative number of pixels in each sample and two columns.}
\usage{
getPixelsCoordinates(data, sample = 1)
}
\arguments{
\item{data}{-> a list from getPeakMatriz()}

\item{sample}{-> just download the pixels from this sample.
if sample < 1, all sample coordinates are returned}
}
\value{
a list:
pixelSize: lateral resolution of pixels
pxCoordinates: matrix with the coordinates (X/Y) of pixels.
}
\description{
returns a matrix with the coordinates of all pixels (X/Y).
If there are multiple samples, they appear sequentially; that is, the matrix has as many rows
as the cumulative number of pixels in each sample and two columns.
}
