% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{nearestIndex}
\alias{nearestIndex}
\title{nearestIndex()
Return the nearest value in data
Successive approximation algorithm.}
\usage{
nearestIndex(value, data)
}
\arguments{
\item{value}{-> reference value}

\item{data}{-> array of sort values}
}
\value{
nearest value index in data to value; -1 if value es out of range
export
}
\description{
nearestIndex()
Return the nearest value in data
Successive approximation algorithm.
}
