% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{peakMatrix}
\alias{peakMatrix}
\title{@name peakMatrix
@title converts the info in the imzML file into a peak matrix.

@param ibdFname:  absolute reference to the file with the ibd extension.
@param imzML:     list with information extracted from the imzML file with import_imzML()
@param params:    specific parameters
"SNR": signal-to-noise ratio
"massResolution": mass resolution with which the spectra were acquired.
"noiseMethod": method for estimating noise.
"minPixelsSupport": minimum percentage of pixels that must support an ion for it to be considered.
"linkedPeaks": two peaks are considered linked if they are closer than the given standard deviation (by defect=3).
@param mzLow:    lower mass to consider
@param mzHigh:   higher mass to consider
@param pxList:   list of pixels. First pixel=1. By default everyone.
@param nThreads: number of threads suggested for parallel processing.
@return lista: peakMatrix, massVector, massResolution, pixelsSupport
peakMatrix: matrix of centroids and the intensity associated with each pixel.
massResolution:
massVector: the mz associated with each column of peakmatrix.
pixelsSupport: number of pixels with intensity >= minPixelsSupport}
\usage{
peakMatrix(ibdFname, imzML, params, mzLow, mzHigh, pxList, nThreads)
}
\description{
@name peakMatrix
@title converts the info in the imzML file into a peak matrix.

@param ibdFname:  absolute reference to the file with the ibd extension.
@param imzML:     list with information extracted from the imzML file with import_imzML()
@param params:    specific parameters
"SNR": signal-to-noise ratio
"massResolution": mass resolution with which the spectra were acquired.
"noiseMethod": method for estimating noise.
"minPixelsSupport": minimum percentage of pixels that must support an ion for it to be considered.
"linkedPeaks": two peaks are considered linked if they are closer than the given standard deviation (by defect=3).
@param mzLow:    lower mass to consider
@param mzHigh:   higher mass to consider
@param pxList:   list of pixels. First pixel=1. By default everyone.
@param nThreads: number of threads suggested for parallel processing.
@return lista: peakMatrix, massVector, massResolution, pixelsSupport
peakMatrix: matrix of centroids and the intensity associated with each pixel.
massResolution:
massVector: the mz associated with each column of peakmatrix.
pixelsSupport: number of pixels with intensity >= minPixelsSupport
}
