% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{remap2ImageCoords}
\alias{remap2ImageCoords}
\title{remap2ImageCoords.}
\usage{
remap2ImageCoords(dataPos)
}
\arguments{
\item{dataPos}{a pos matrix as it is in rMSI data object.

This function should be only used to implement data importers from foreign formats.
This functions maps a MALDI motors coors space to a image coord space.
dataPos matrix is a two columns matrix where first column stores x positions and second y pixel positions.
a remapped dataPos matrix do not contain empty raster positions neighter offsets.}
}
\value{
the dataPos matrix remapped.
export
}
\description{
remap2ImageCoords.
}
