% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rPeakMatrix.R
\name{rGetIntensityFromFile()}
\alias{rGetIntensityFromFile()}
\alias{getIntensityFromFile}
\title{returns a column with pixel intensities from the peak matrix:}
\usage{
getIntensityFromFile(file, column)
}
\arguments{
\item{file}{-> file name with peak matrix (tmpPeakMatrix.bin)}

\item{column}{-> desired column (first = 1)}
}
\value{
a vector with intensity of each pixel, m/z, mass resolution and number of pixels with non-zero magnitude.
}
\description{
returns a column with pixel intensities from the peak matrix:
}
