% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{rGetIntensityMatrix()}
\alias{rGetIntensityMatrix()}
\alias{getIntensityMatrix}
\title{Returns a list containing the intensity matrices associated with each sample.}
\usage{
getIntensityMatrix(data, sampleList = c(0))
}
\arguments{
\item{data}{-> a list from getPeakMatriz()}

\item{sampleList->}{list of samples (numerical values). By defect or zero, return all samples together}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     -> a list of intensity matrices: row = pixels; column=centroids
}\if{html}{\out{</div>}}
}
\description{
Returns a list containing the intensity matrices associated with each sample.
}
