% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{rGetMatrix()}
\alias{rGetMatrix()}
\alias{getMatrix}
\title{returns the intensity matrix associated with a given sample.}
\usage{
getMatrix(data, sample = 1)
}
\arguments{
\item{data}{-> a list from getPeakMatriz()}

\item{sample}{-> just download the matrix intensity from this sample.}
}
\value{
a matrix -> row = pixels; column=centroids
}
\description{
returns the intensity matrix associated with a given sample.
}
