% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotGaussianSpectrum2}
\alias{rPlotGaussianSpectrum2}
\title{Presents an image with the Gaussians of two spectrum fragments,
along with the curve resulting from the sum of them (in red).
The blue circles indicate the original concentration data to be adjusted.
The drawInfo parameters come from getPixelGaussians() or getGaussiansFromSpectrum()}
\usage{
rPlotGaussianSpectrum2(drawInfo1, drawInfo2, minMass = 0, maxMass = 0)
}
\arguments{
\item{drawInfo$gaussians:}{matrix
column 1 -> mean values of each Gaussian.
column 2 -> standard deviation.
column 3 -> factor associated with each Gaussian.}

\item{drawInfo$mass:}{vector (in Daltons).}

\item{drawInfo$intensity:}{vector of intensities}

\item{minMass:}{minimun mz}

\item{maxMass:}{maximun mz}
}
\value{
-1 if KO; 0 if OK
}
\description{
Presents an image with the Gaussians of two spectrum fragments,
along with the curve resulting from the sum of them (in red).
The blue circles indicate the original concentration data to be adjusted.
The drawInfo parameters come from getPixelGaussians() or getGaussiansFromSpectrum()
}
