% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotGaussianSpectrum3}
\alias{rPlotGaussianSpectrum3}
\title{Presents an image with the Gaussians of three spectrum fragments,
along with the curve resulting from the sum of them (in red).
The blue circles indicate the original intensity data to be adjusted.
The drawInfo parameters come from getPixelGaussians() or getGaussiansFromSpectrum()}
\usage{
rPlotGaussianSpectrum3(
  drawInfo1,
  drawInfo2,
  drawInfo3,
  minMass = 0,
  maxMass = 0,
  rMSI2_peaks1 = -1,
  rMSI2_peaks2 = -1,
  rMSI2_peaks3 = -1
)
}
\arguments{
\item{drawInfo$gaussians:}{matrix
column 1 -> mean values of each Gaussian.
column 2 -> standard deviation.
column 3 -> factor associated with each Gaussian.}

\item{drawInfo$mass:}{vector (in Daltons).}

\item{drawInfo$intensity:}{vector of intensities}

\item{minMass:}{minimun mz}

\item{maxMass:}{maximun mz}

\item{rMSI2_peaks1:}{median resolution peaks from rMSI peak matrix}

\item{rMSI3_peaks1:}{high   resolution peaks from rMSI peak matrix}
}
\value{
-1 if KO; 0 if OK
}
\description{
Presents an image with the Gaussians of three spectrum fragments,
along with the curve resulting from the sum of them (in red).
The blue circles indicate the original intensity data to be adjusted.
The drawInfo parameters come from getPixelGaussians() or getGaussiansFromSpectrum()
}
