% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotIon}
\alias{rPlotIon}
\title{visualize an ion.}
\usage{
rPlotIon(pMatrix, mz)
}
\arguments{
\item{pMatrix:}{peak matrix from getPeakMatrix().}

\item{mz:}{desired mass (Da).}
}
\description{
visualize an ion.
}
