% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotSpectrum}
\alias{rPlotSpectrum}
\title{Presents an image with the info of two spectrum fragments,}
\usage{
rPlotSpectrum(values, mzLow, mzHigh)
}
\arguments{
\item{mzLow}{: minimun mz}

\item{List:}{values come from getAverageSpectrum() or getAverageGaussianSpectrum()
values$averageMz: vector of mases(in Daltons).
values$averageIntensity: vector of intesities}

\item{mzHigh:}{maximun mz}
}
\description{
Presents an image with the info of two spectrum fragments,
}
