% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{saveContext}
\alias{saveContext}
\title{save info about conditions of test}
\usage{
saveContext(samples, outBaseDir, params)
}
\arguments{
\item{samples}{-> a character vector with all raw files}

\item{outBaseDir}{-> output directory where files will be created}
}
\value{
none
}
\description{
save info about conditions of test
}
