% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getAverageSpectrum}
\alias{getAverageSpectrum}
\title{It obtains the average value of the intensities from all data into .imzML file.
Noise is not taken into account.}
\usage{
getAverageSpectrum(
  data_file,
  initMass = 0,
  finalMass = 0,
  pxList = c(0),
  massResolution = 30000,
  overSampling = 4,
  imzMLChecksum = F,
  fixBrokenUUID = F
)
}
\arguments{
\item{data_file:}{absolute reference to the file with the imzML extension.}

\item{initMass:}{lower mass to consider}

\item{finalMass:}{higher mass to consider}

\item{pxList:}{list of pixels. First pixel=1. By default everyone.}

\item{massResolution:}{mass resolution with which the spectra were acquired (mz/deltaMz).}

\item{overSampling:}{interval between points on the mass axis = massResolution/overSampling.}

\item{imzMLChecksum:}{if the binary file checksum must be verified, it can be disabled for convenice with really big files.}

\item{fixBrokenUUID:}{set to FALSE by default to automatically fix an uuid mismatch between the ibd and the imzML files (a warning message will be raised).}
}
\value{
a list: averageMz and averageIntensity
averageMz: array of masses at intervals of the resolution/overSampling
averageIntensity: array of average Gaussians values
}
\description{
It obtains the average value of the intensities from all data into .imzML file.
Noise is not taken into account.
}
