% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getAverageSpectrum}
\alias{getAverageSpectrum}
\title{It obtains the average value of the intensities from all data into .imzML file.
Noise is not taken into account.}
\usage{
getAverageSpectrum(dataFile, params)
}
\arguments{
\item{dataFile:}{absolute reference to the file with the imzML extension.}

\item{parmas:}{initMass: initial  mass to consider. By default, the minimum value from the entire range of masses is used.
finalMass: final    mass to consider. By default, the maximum value from the entire range of masses is used.
pxList: list of pixels. First pixel=1. By default everyone.
overSampling: interval between points on the mass axis = tolerance/overSampling.
imzMLChecksum: if the binary file checksum must be verified, it can be disabled for convenice with really big files.
fixBrokenUUID: set to FALSE by default to automatically fix an uuid mismatch between the ibd and the imzML files (a warning message will be raised).}
}
\value{
a list: averageMz and averageIntensity
averageMz: array of masses at intervals of the tolerance/overSampling
averageIntensity: array of average Gaussians values
}
\description{
It obtains the average value of the intensities from all data into .imzML file.
Noise is not taken into account.
}
