% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getColumnFromFile()}
\alias{getColumnFromFile()}
\alias{getColumnFromFile}
\title{returns a column information of the peak matrix:}
\usage{
getColumnFromFile(file, mass, sample)
}
\arguments{
\item{file}{-> file name with peak matrix (tmpPeakMatrix.bin)}

\item{mass}{-> reference to the desired initial column of the peak matrix (Da).}

\item{sample}{-> just download the pixels from this sample.
if sample < 0, all sample coordinates are returned}
}
\value{
a list:
intensity: vector of intesities
mass: mass associated with the column of peakMatrix.
massResolution: final mass resolution at centroid.
pixelsSupport: number of pixels in column with non-zero intensity.
}
\description{
returns a column information of the peak matrix:
}
