% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getGaussiansFromSpectrum}
\alias{getGaussiansFromSpectrum}
\title{Form Gaussians over each peak of the given spectrum.}
\usage{
getGaussiansFromSpectrum(intensity, mass, params)
}
\arguments{
\item{Intensity:}{vector with value associated with each mass point.}

\item{mass:}{vector with mass/charge information}

\item{params:}{initMass:    initial  mass to consider. By default, the minimum value from the entire range of masses is used.
finalMass:   final    mass to consider. By default, the maximum value from the entire range of masses is used.
SNR:         signal-to-noise ratio
noiseMethod: method for estimating noise (estnoise_diff, estnoise_sd, estnoise_mad).}
}
\value{
List: gaussians, mass intensity, SNR, noise
gaussians: matrix with parameters for each Gaussian
mass: vector with the masses of the raw spectrum
intensity: intensity associated with each mass of the raw spectrum
SNR: signal-to-noise ratio associated with each mass of the raw spectrum.
noise: noise estimation
}
\description{
Form Gaussians over each peak of the given spectrum.
}
