% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getPeakMatrix}
\alias{getPeakMatrix}
\title{obtains the peak matrix from imzML files.}
\usage{
getPeakMatrix(dataFiles, params, outDirectory)
}
\arguments{
\item{params}{"tolerance": desired tolerance for the centroids (ppm)
"SNR": signal-to-noise ratio (by defect=1)
"noiseMethod": method for estimating noise (by defect="estnoise_mad").
"minPixelsSupport": minimum percentage of pixels that must support an ion for it to be considered (by defect=1).
initMass: initial mass to consider. By default, the minimum value from the entire range of masses is used.
finalMass: final   mass to consider. By default, the maximum value from the entire range of masses is used.
pxList: list of pixels. First pixel=1. By default everyone.
peakMethod: algorithm for shaping the peaks: uGaussians (defect), iGaussians, derivative (u=united, i=isolated)
intMethod: intensity values for the binning stage: mean (defect), max
nThreads: number of threads for parallel processing (by default maxCores-1)
imzMLChecksum: if the binary file checksum must be verified, it can be disabled for convenience with really big files.
fixBrokenUUID: set to FALSE by default to automatically fix an uuid mismatch between the ibd and the imzML files (a warning message will be raised).}

\item{dataFiles:}{A list of the absolute paths of the files to be processed is required.
Alternatively, a single filename with the 'txt' extension containing the absolute paths of all files to be processed is acceptable.
In file.txt, lines are considered comments if they begin with the '#' character or the blank character.
Only files in 'imzML' format are recognized.
The attached binary file, with the 'ibd' extension, must be in the same directory that 'imzML' files.}

\item{outDirectory:}{Absolute path to the directory where the support files will be created.}
}
\value{
a list with the input parameters
}
\description{
obtains the peak matrix from imzML files.
}
