% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getPxGaussians}
\alias{getPxGaussians}
\alias{getPixelGaussians}
\title{It retrieves information from a single pixel.}
\usage{
getPixelGaussians(
  data_file,
  params = c(),
  initMass = 0,
  finalMass = 0,
  pixel = 1,
  imzMLChecksum = F,
  fixBrokenUUID = F
)
}
\arguments{
\item{params}{"tolerance": desired tolerance for the centroids (ppm)
"SNR": signal-to-noise ratio
"noiseMethod": method for estimating noise.}

\item{data_file:}{absolute reference to the file with the imzML extension.}

\item{initMass:}{Initial  mass to consider. By default, the minimum value from the entire range of masses is used.}

\item{finalMass:}{Final    mass to consider. By default, the maximum value from the entire range of masses is used.}

\item{pixel:}{spectrum to evaluate. First pixel=1}

\item{imzMLChecksum:}{if the binary file checksum must be verified, it can be disabled for convenice with really big files.}

\item{fixBrokenUUID:}{set to FALSE by default to automatically fix an uuid mismatch between the ibd and the imzML files (a warning message will be raised).}
}
\value{
a list:
gaussians: parameters of each Gaussian.
mass: Vector with raw masses.
intensity: vector with raw intensities;
SNR: signal-to-noise ratio for each point of mass.
noise: noise estimation.
}
\description{
It retrieves information from a single pixel.
}
