% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rGetCoordinatesFromFile()}
\alias{rGetCoordinatesFromFile()}
\alias{rGetCoordinatesFromFile}
\title{returns a matrix with the coordinates of all pixels (X/Y).
If there are multiple samples, they appear sequentially; that is, the matrix has as many rows
as the cumulative number of pixels in each sample and two columns.}
\usage{
rGetCoordinatesFromFile(fileName)
}
\arguments{
\item{file}{-> file name with peak matrix}
}
\value{
a matrix with the coordinates (X/Y) of all pixels.
}
\description{
returns a matrix with the coordinates of all pixels (X/Y).
If there are multiple samples, they appear sequentially; that is, the matrix has as many rows
as the cumulative number of pixels in each sample and two columns.
}
