% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rGetCoordinatesFromFile()}
\alias{rGetCoordinatesFromFile()}
\alias{rGetPixelsCoordinates}
\title{returns a matrix with the coordinates of all pixels (X/Y).
If there are multiple samples, they appear sequentially; that is, the matrix has as many rows
as the cumulative number of pixels in each sample and two columns.}
\usage{
rGetPixelsCoordinates(fileName, sample)
}
\arguments{
\item{sample}{-> just download the pixels from this sample.
if sample < 1, all sample coordinates are returned}

\item{file}{-> file name with pixels coordinates (_pixelsCoord.bin)}
}
\value{
a matrix with the coordinates (X/Y) of pixels.
}
\description{
returns a matrix with the coordinates of all pixels (X/Y).
If there are multiple samples, they appear sequentially; that is, the matrix has as many rows
as the cumulative number of pixels in each sample and two columns.
}
