% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rPeakMatrix.R
\name{rGetMetaDataFromFile}
\alias{rGetMetaDataFromFile}
\alias{getMetaDataFromFile}
\title{returns generic information about the peak matrix located in a file.}
\usage{
getMetaDataFromFile(file)
}
\arguments{
\item{file}{-> file name with peak matrix  (tmpPeakMatrix.bin)}
}
\value{
a list:
nSamples     -> number of samples analyzed.
totalPx      -> total number of pixels (cumulative of each sample).
nIons        -> number of columns in the matrix.
pixelsSample -> vector with the pixels in each sample.

a list:
nSamples     -> number of samples analyzed.
totalPx      -> total number of pixels (cumulative of each sample).
nIons        -> number of columns in the matrix.
pixelsSample -> vector with the pixels in each sample.
}
\description{
returns generic information about the peak matrix located in a file.

returns generic information about the peak matrix located in a file.
}
