% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotGaussianSpectrum}
\alias{rPlotGaussianSpectrum}
\title{Presents an image with the Gaussians from a spectra fragment,
along with the curve resulting from the sum of all of them (in red).
The blue circles indicate the original intensity data to be adjusted.
The drawInfo parameters come from getPixelGaussians() or getGaussiansFromSpectrum()}
\usage{
rPlotGaussianSpectrum(drawInfo, minMass = 0, maxMass = 0, sum = TRUE)
}
\arguments{
\item{drawInfo$gaussians:}{matrix
column 1 -> mean values of each Gaussian.
column 2 -> standard deviation.
column 3 -> factor associated with each Gaussian.}

\item{drawInfo$mass:}{vector (in Daltons).}

\item{drawInfo$intensity:}{vector of intensities. It can be null or non-existent.}

\item{sum:}{if TRUE, the sum of Gaussian is displayed.}
}
\value{
-1 if KO; 0 if OK
}
\description{
Presents an image with the Gaussians from a spectra fragment,
along with the curve resulting from the sum of all of them (in red).
The blue circles indicate the original intensity data to be adjusted.
The drawInfo parameters come from getPixelGaussians() or getGaussiansFromSpectrum()
}
