% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotIon}
\alias{rPlotIon}
\title{visualize an ion.}
\usage{
rPlotIon(intensity, coordinates)
}
\arguments{
\item{intensity}{:intensity values for each coordinates.}

\item{coordinates:}{position en plane X/Y.}
}
\description{
visualize an ion.
}
